// packages/Morado/src/store/components/Product/types.ts
type ICartButton = {
  text?: string;
  style?: string;
  product: Product;
  addToCart: () => void;
  disabled?:boolean
};